requirements.txt
python-dotenv
flask
flask-smorest
marshmallow
pymongo
